Refactor code after PR review
